Refactor into utils, rather than command
